Add perfect capture and return for match
optional<T>::match should be able to capture the arguments by "universal
reference". This way, the callables can mutate their internal
state and the user does not encounter the surprise of mutating a
temporary copy.

The return value of optional<T>::match should be exactly the same as the
(common) return value of the two callables.